Clean up iOS chat keyboard review findings
