+++ Packages/iOS/CmuxAgentChatUI/Sources/CmuxAgentChatUI/Composer/ChatAccessoryShortcutSemanticAction.swift
@@ -1,0 +1,7 @@
+/// Composer-owned behavior attached to a host-provided visual shortcut.
+public enum ChatAccessoryShortcutSemanticAction {
+    /// Paste clipboard contents into the chat composer.
+    case paste
+    /// Dismiss the chat composer's keyboard.
+    case dismissKeyboard
+}
